Remove locking around stats and use __sync insts instead

+++ tpool.h
@@ -1,18 +1,19 @@
 #ifndef TPOOL_H
 #define TPOOL_H
 
+#include <stdint.h>
 #include "spinlock.h"
 #include "kqueue.h"
 
 struct tpool_stats {
   int active_threads;
   int blocked_threads;
-  double active_threads_sum;
-  double blocked_threads_sum;
+  uint64_t active_threads_sum;
+  uint64_t blocked_threads_sum;
   int active_threads_samples;
   int blocked_threads_samples;
   int items_processed;
-  double processing_time_sum;
+  uint64_t processing_time_sum;
 };
 
 struct tpool {
@@ -36,9 +37,9 @@ struct tpool_stats tpool_get_stats(struct tpool *t);
 int tpool_get_items_processed(struct tpool_stats *prev,
                               struct tpool_stats *curr);
 double tpool_get_average_active_threads(struct tpool_stats *prev,
-                                        struct tpool_stats *curr);
+                                          struct tpool_stats *curr);
 double tpool_get_average_blocked_threads(struct tpool_stats *prev,
-                                         struct tpool_stats *curr);
+                                           struct tpool_stats *curr);
 double tpool_get_average_processing_time(struct tpool_stats *prev,
                                          struct tpool_stats *curr);
 void tpool_print_items_processed(char *prefix,
